Working on sprint, training course manager and targets
Sprint is still not working based on the input interaction type (hold or tap).

Created a training course manager script and corresponding game object in the hierarchy.

Created a U.I. canvas in the hierarchy and added the following game objects: tutorial text placeholder, target text, current score text and bullets remaining text.

Created a target script and corresponding game object.

Created a placeholder enemy game object.

Updated the gunplay manager to remove the commented out sections for input interactions and added a boolean to check if the gun is currently being used in a training course.