Basic Enemy A.I.
Created a basic enemy A.I. prefab and created several variations of this enemy A.I. for the different enemy types.

Updated the training course system so it now concludes the training.

Created an asymmetrical gameplay manager which instantiates the enemy A.I. and manages the gameplay when the player finishes training.