Further Fixes, Tweaks and Small Features
Added a sound manager to the scene hierarchy.
Added functionality in the gunplay manager script for two SFX clips, one for tapping fire and the other for holding fire.
Made changes to the excluded layers on the sphere collider for the projectile prefab.
Turned the target highlight column into a prefab.
Finished implementing the target highlight column into the training course manager.