Modifying Scripts
Made adjustments to the included and excluded layers of the colliders in all the enemy prefabs.

Modified the game manager script to include events for when the player is hit or killed and for when an enemy is hit.

Connected the new events in the game manager to the player script, gunplay manager script and enemy script.

Moved one of the walls in the fps test scene to close an unintended gap.