Continuing Progress
Made changes to the InputManager script to allow interacting with the pause menu when playing in the FPS test scene.

Updated the gunplay manager and projectile scripts to use events instead of colliders when collisions occur.

Added a boolean for showing the attack range and another boolean for showing the sight range in the enemy A.I. controller script.

Added debugging region to the update function of the enemy A.I. controller script.

Also added two debug.log statements for when the enemy A.I. is hit and when it is killed.

Added two public booleans to the game manager for the training course game state and the FPS game state.

Added the player killed event call in the first person movement script so to disable the player movement when the player is killed.

Commented out the input manager reference variable and reference in the awake function of the Cinemachine POV extension script.

Increased the projectile life span from 5 to 10 in the projectile script.

Removed the body and head collider references in the player game character script.

Set the layer to enemy on the enemy head child object for each of the enemy prefabs.

Fixed the on click events for all the pause menu buttons in the FPS_Mechanics Test scene.

Updated the input actions in the events system U.I. input module component for the FPS_Mechanics test scene.

Still need to fix shooting and killing enemies.