Updated FPS Gameplay
Made multiple tweaks to the enemyAiController script such as lowering the default health.

Changed each enemy's head from a squashed cube to a sphere and set 'enemy' in the excluded layers for the head collider to prevent collision from occurring between the head and the body.

Enabled the 'is kinematic' setting for each enemy's rigidbody.

Enemies can now be properly killed by the player's gun.

Gave enemies glasses and created a new basic purple material for these.

Changed the level design of the FPS_Mechanics Test Scene and disabled kinematics on each wall's rigidbody. Also enabled 'convex' and 'provides contacts' on each wall's collider.

Updated the pause functionality in the InputManager so that both the escape key and the resume button can be used to resume the game when paused.

Added a new headshot damage variable to the GunplayManager script.

Still need to fix player movement being locked after resuming using the escape key.

Removed unused usings/name spaces in the Target script.