Creating Two Characters At Once
Duplicated the FPS_Mechanics Test scene and renamed the duplicate to 'Two Characters At Once Test'.

Disabled the enemies and navmesh in the new scene.

Duplicated the main camera of the scene and renamed each to Player 'X' Camera.

Created layers for player 1 and a player 2 and set each player character to their respective layers.

Duplicated the canvas for the second character and renamed each to their respective players.

[BUG] Player 1's gun currently rotates while moving the camera.

Duplicated the Cinemachine virtual camera and renamed both to Player 'X' - First Person Camera.