First update
Created a low friction physics material and applied it to the collider component of the body child object of the player character.

Duplicated the level_design_test scene to allow for level design changes specifically for this prototype. Renamed this duplicated scene to Prototype_1_Level and removed all game objects pertaining to the two characters at once prototype.

Created a tutorial U.I. screen with start game and quit game buttons. Quit button references the OnApplicationQuit function from the GameManager script.

Created a StartGame script in the GameManager script and a variable to reference the tutorial screen.

Modified the pause and game over screens to remove unnecessary components such as quit to main menu.

Commented out references to the Quit Prompt Screen and the Return to Pause Screen and Quit to Main Menu buttons in the GameplayManager script.

Renamed the 'bullets remaining' text elements in (hierarchy and in game) to Ammo.

Created a health title text element and a health text element. Moved both of these to be next to the ammo title and ammo text elements.

Updated the playerGameCharacter script to reference the new health text element. Created colour variables for the text, so that the colour corresponds to the amount of health remaining. This still needs to be tested.

Created a variant of the default TMP font (LiberationSans SDF) to apply the extra font settings without affecting all other text in the project.

Commented out all references to the two characters at once prototype in the InputManager script.

Added recoil functions to the GunplayManager script. These still need to be properly tested.

Set the move speed and sprint speed to zero in the inspector window of the FirstPersonMovement script component on the player game object.

Commented out the character movement code in the FirstPersonMovement script.