Update 2
Created a new material for the default TextMeshPro font - Liberation SDF.

Added title text (which says "Paused") and a settings button to the pause screen.

Created the OnSettingsButtonClicked function for this button in the GameManager.

Created the settings page. Added all the U.I. elements for the settings page and setup basic layout for these settings.

Removed the isInFPS boolean from the GameManager script as it's no longer relevant.

Created the settings U.I. page with a return button, two sliders for mouseX and mouseY sensitivity and two toggle U.I. objects: one for invert mouse Y and the other for mouse acceleration.

Created functions for the settings page U.I. objects in the InputManager.

Add references to all the new U.I. objects to the GameManager.

Updated the pause and resume functions in the GameManager to enable and disable the new settings U.I. elements.

Set outer walls height (y axis) to 5.

Started trying to implement gun recoil moving the player character.