Update 3
Disabled the character controller component on the player game object.

Added a rigidbody to the player game object.

Adjusted the animation curves for the gun recoil.

Removed the isFPSTesting boolean from the GunplayManager script as it was no longer relevant and causing the gun to start disabled.

Removed a missed function for training courses in the GunplayManager script.

Added a reference in the FirstPersonMovement script to the collider component of the body child object.

Set the jump height to zero in the inspector for the FirstPersonMovement script.

Finished the simple level design for the first prototype.

Added enemies to the level for the first prototype. Then modified these enemies.

Cleared the navmesh and rebaked it.

Both the level and the enemies needs to be tested and then adjusted accordingly.