Update Four
Made the projectile prefab's model size half the previous size.

Added wall and ground layers to  the included layers to detect for the projectile's sphere collider.

Added a rigidbody and a capsule collider to the player character object and changed the if conditions in the update function of the FirstPersonMovement to stop player movement when colliding with walls.

Created a behaviour graph and a behaviour blackboard to start learning behaviour graphing so to improve the enemy A.I.

Deleted the CameraShake script as it was no longer necessary.