Update 5
Resized the capsule collider on the player game object to better fit the body capsule.

Fixed the health text variable not updating from the playerGameCharacter script.

Fixed the tutorial screen not showing at the start of the prototype.

Created a start function for the GameManager script to start the scene with the tutorial and the U.I. action map enabled.

Fixed the player character getting stuck mid-air.

Modified the DisableGameInput function in the InputManager to make the cursor visible and to confine it to the game window.