Update 7
Updated the tutorial text screen to fix the grammatical error in the second line. Also added a line to inform the player that they can change settings from the pause menu.

Fixed the health text not updating in the playerGameCharacter script. Did this by moving the health text into the update function.

Fixed the bug where the player couldn't shoot after pressing the restart button in the pause menu. This was fixed by changing the code to load the scene by name, not by getting the current scene.

Replaced the two restart level functions from the pause and game over screens with references to one restart scene function.

Added a LoadMainMenu function to the GameManager to also reduce the number of repetitive functions within the GameManager script.

Fixed the bug where the game over screen was not interactable. This was partly due to a mistake in the code in which the event didn't call properly.

Commented out the isPlayerDead if statement in DisableGameInput function of the InputManager as it was preventing player interaction with the game over screen.

Fixed the bug where the reloading text overflowed.

Enabled the trigger option for the projectile prefab's sphere collider. This is partly of the fix to prevent projectiles from going through walls. Also unchecked the Provides Contacts option as it was unnecessary.

Changed the OnCollisionEnter function in the projectile script to OnTriggerEnter.

Set the level priority to 1 for all the level geometry objects' colliders.

Removed the timerfinished boolean from the Projectile script.