New Features
Duplicated the Prototype_1_Level scene and named it to Arena_Level.

Created a text element called "Reload Prompt" in the Arena_Level scene.

Created variables in the GameManager to reference the reload prompt text and to create a timer for showing or hiding it.

Created ShowReloadPrompt and HideReloadPrompt functions in the GameplayManager script.

Created an update function in the GameplayManager script for the reload prompt timer.

Created a skip tutorial boolean in the GameManager script.

Modified the start function of the GameManager to allow for the new skip tutorial boolean.

Renamed the existing player prefab to "Player (old) and created a new player prefab using the player from the imported first prototype scene. Named the new player prefab "Player (new)".

Created text components for mouseXSensitivity and mouseYSensitivity and placed them next to their respective sliders on the settings page of the pause menu.

Starting trying to fix the bug where the bullet collision decals are left floating in the air.

Starting trying to fix a bug where the muzzle flash VFX doesn't show.