Renaming
Renamed all the prefabs and scripts for this prototype, so that they can be merged with the main branch later on without causing any merge conflicts.

Deleted the testGun script as it is not necessary for the first prototype.

Fixed the empty CompareTag string condition in the OnTriggerEnter function of the projectileScript.