Recreating Assets
Recreated the following prefabs to avoid merge conflicts:
Projectile
Gun

Recreated the following scripts to avoid merge conflicts:
CinemachineShake
CinemachinePOVExtension
enemyAiController
GameManager
InputManager
GunplayManager
FirstPersonMovement
playerGameCharacter
projectileScript

Fixed all the variables for all the script components and all the prefabs and game objects that reference these scripts.

Fixed the missing InputManager function references for both the InvertMouseY and the MouseAcceleration toggles on the settings screen.

Unchecked the invert setting for the vertical axis of the aim subcomponent of the cinemachine virtual camera component on the first person camera.