Small Fixes
Reverted the amplitude gain and the frequency gain of the cinemachine virtual camera noise to 0.

Fixed the missing InputManager function references for both the InvertMouseY and the MouseAcceleration toggles on the settings screen.

Fixed the missing CompareTag condition in the

Unchecked the invert setting for the vertical axis of the aim subcomponent of the cinemachine virtual camera component on the first person camera.