Minor Features and Fixes
Added a field of view (FOV) slider and corresponding text to the settings menu.

Added functionality for the field of view slider and corresponding text to the InputManager script.

Added a toggle for the reload prompt text to the settings menu.

Fixed the mouse acceleration and invert mouse toggles elements in the settings page by making the corresponding booleans for both settings to {get; set;}.

Added functionality for the reload prompt toggle to the GameManager script by changing the reloadprompt to public and adding {get; set;}.

Fixed the bug where bullet decals would incorrectly spawn, causing them to float mid-air.

Fixed a performance bug where muzzle flashes would stay in the scene hierarchy infinitely after being instantiated by GunplayManager. This was fixed by setting the stop action to destroy for the MuzzleFlash01 prefab.

Removed commented out code from the GunplayManager script which will no longer be needed.

Improved the feel of the jump by adjusting the player's jump speed and time to jump apex variables.