Pause Menu Minor Features
Added a text element to the pause screen to show the level objective.

Added a tutorial button to the pause menu.

Added a return button on the tutorial screen.

Added references for the tutorial button, return from tutorial screen button, objective text elements to the GameManager script.

Added functions to the GameManager for updating the objective text, showing the tutorial screen, returning from the tutorial screen, hiding the return button, showing the tutorial screen and hiding the tutorial screen.

Added a variable for the scene name to the GameManager script and changed the OnRestartButtonPressed and RestartLevelFromGameOverScreen functions to include this new variable instead of the previous in code string.

Added a return button to the tutorial screen and disabled it.