Second update
Created CameraID script to attach to each player character's camera.

Populated the P2_CameraManager with some code for cycling through characters, but this still needs to be tested.

Created the P2_PlayerCharacterBase script and populated it with some code to correspond to the events for changing characters and when a character is killed in the P2_GameManager script.

Added a new action called "Change Character" to the PlayerControls inputactions asset.

Added a corresponding boolean for the Change Character input action to the P2_InputManager script.

Added an event to the P2_GameManager script for when the player is killed.