Third Update
Created a scene called Prototype_2_Level.

Populated the Prototype_2_Level with game objects from the Arena_Level_Scene.

Created game objects PlayerManager, CameraManager and Players in the Prototype_2_Level scene.

Created a script called P2_ProjectileScript and populated it with code copied from the original ProjectileScript.

Updated the P2_PlayerCharacterBase script with code copied over from the playerGameCharacter script.

Updated the P2_GameManager, to reference the P2_InputManager instead of InputManager.

Updated the GameManager reference to P2_GameManager in the following scripts: P2_ProjectileScript, P2_PlayerManager and P2_InputManager.

Created a script called P2_GunplayManager and copied over the code from the GunplayManager script. Updated all references to the InputManager and GameManager to include the prefix "P2_" for the second prototype versions of those scripts.

Replaced all instances of Invoke() in the P2_GunplayManager with respective timers.

Created a folder in Prefabs called Prototype_2.

Recreated the projectile prefab and renamed it P2_Projectile.

Created duplicated the existing player character and cinemachine virtual camera in the scene.

Fixed all missing references within the scene for the InputManager, GameManager, PlayerManager, CameraManager, Player and Gun game objects.

Created a script called P2_fpsMovement and copied over the code from the FirstPersonMovement script. Updated all references to the GameManager and InputManager to include the prefix "P2_" for the second prototype versions of those scripts.

Duplicated the existing PlayerControls InputAction asset and renamed it to P2_PlayerControls.

Updated reference to the input action asset in the P2_InputManager script.

Created a script called P2_CinemachineShake and copied over the code from the CinemachineShake script. Updated the only reference to the GameManager to "P2_GameManager".

Replaced the CinemachineShake script for both player character cinemachine virtual cameras with the P2_Cinemachine script.

Created a script called P2_enemyAttack and copied over the code from the enemyAttack script. Updated the only reference to the GameManager to "P2_GameManager".

Created a folder inside the Behaviour folder called Prototype 2. Duplicated the Enemy Behaviour Graph and renamed the duplicate to "P2_Enemy Behaviour Graph".

Replaced the variable to reference the enemyAttack script in the P2_Enemy Behaviour graph with a variable to reference the P2_enemy attack script.

Unpacked the Enemy_Basic prefab in the Prototype_2_Level scene and renamed it to P2_Enemy_Basic.

Replaced the enemyAttack script on P2_Enemy_Basic with the P2_enemyAttack script.

Updated the Behaviour Agent component on P2_Enemy_Basic to reference the P2_Enemy Behaviour Graph.