Fifth Update
Created an empty game object in hierarchy, called it "Heist Payload" and added a cube as a child. Renamed the cube child object to "Model".

Set the material of the cube child object to a gold material.

Added a rigidbody to the model child object  and set its mass to 7.

Set the mass in the player 1 rigidbody to: 5.

Set the mass in the player 2 rigidbody to: 8.5.

Set the mass in the player 3 rigidbody to: 2.

Added a reference to the heist payload game object in the playerManager script.

Created a new Physics material, called Friction. Applied this physics material on the box collider component of the model child object of Heist payload.

Created a script called P2_enemyHealth and copied over code from the enemyHealth script.

Replaced the reference to the GameManager script in the enemyHealth with a reference to the P2_GameManager script.

Replace the enemyAttack and enemyHealth scripts on the P2_Enemy_Basic prefab with the P2_enemyAttack and P2_enemyHealth scripts.

Added a configurable joint component to all player characters and set the connected body variable to the rigidbody of the Heist Payload child object.