Sixth Update
Implemented functionality for pushing other game objects in the P2_CharacterBase script.

Created a new layer called Pushable, for objects that can be pushed.

Created a script called P2_RangeSensor and copied over code from the RangeSensor script.

Replaced the reference to the GameManager script in the RangeSensor script with a reference to the P2_GameManager script.

Replaced the RangeSensor script component of P2_Enemy_Basic prefab with the P2RangeSensor script.

Created a custom action for P2_EnemyBehaviourGraph called P2RangeSensorAction copied over code from the RangeSensorAction script.

Created a custom condition for P2_EnemyBehaviourGraph called P2_PlayerSighted and copied code from the PlayerSighted script.

Updated P2_EnemyBehaviourGraph to replace the RangeSensor action with the P2_RangeSensor action.

Updated the P2_EnemyBehaviourGraph to replace the condition of the first branch with the P2_PlayerSighted condition.

Created a script called P2_PlayerDetector and attached it as a component of the Heist Payload game object.