Ninth Update
Fixed all the broken references for all the interactive U.I. elements in the Prototype_2_Level scene.

Increased the size of the objective text on the pause screen.

Updated the objective text on the pause screen.

Updated the player character transforms.

Fixed a bug where the FOV slider in the settings screen would allow mouse movement.