Eleventh Update
Updated the friction physics material dynamic and static friction values.

Updated the freeze position and rotation on the rigidbody on the Heist Payload game object.

Added Environment as an included layer on the box collider of the Heist Payload game object.

Updated the layer to "Gun" for all the player character guns.

Created a new tag called "Pushable" and assigned it to the Heist Payload game object.

Updated the collision in the P2_Projectile script to delete the projectile if it collides with the pushable layer (i.e. the heist payload game object).

Implemented a layermask called "isPushable" in the GunplayManager script and updated the raycast in the FireGun function to provide visual feedback if it hits the heistPayload.

Duplicated the HUD Health Bar child of the canvas and renamed it to HUD Ammo Bar.

Created a script called P2_AmmoHudBar and copied over the code from the P2_HealthBar script. Attached the script to the new HUD Ammo Bar U.I. object.

Disabled the interactable setting in the slider component for both the HUD Health Bar and HUD Ammo Bar.

Added a new TextMeshPro text object to the P2_Hackable_Door prefab and named it Back Text.

Renamed the Door Text of the P2_Hackable_Door prefab to Front Text.

Updated the P2_HackableDoor script to reference the new TextMeshPro text object.

Set the tags and layers for all of the following level geometry objects to environment: desks, chairs and pillars.

Set the tags and layers for all of the following level geometry objects to environment: shorter width walls,  shorter height walls, inner walls, regular doorways and closed doors.

Fixed the minor bug where the hacking progress bar did not match the hacking time interval.

Changed the OnTriggerEnter function in the P2_PlayerDetector script to OnTriggerStay. This should hopefully fix any bugs with script incorrectly detecting player characters are out of range when standing right next to it.

Added a controls button to the pause screen object under the canvas.

Created a child object called Controls Page to the pause screen object under the canvas.

Created a new tag called "PlayerOther" and assigned it to each player character's Body and Head child objects.

Updated the P2_ProjectileScript to include the new tag when colliding the player characters.