Fifthteenth Update
Created a new text object under the Game Over screen called Level Completed Text.

Adjusted the navigation and attack distance variables of behaviour agent component for each P2_Enemy_Basic prefab in the Prototype_2_Level scene.

Changed the sphere collider radius to 20 on the P2_Enemy_Basic prefab.

Removed the detectionMask variable from the P2_RangeSensor script.

Added a healthbar to the canvas child of the P2_Enemy_Basic prefab.

Added a P2_HealthBar script to the P2_Enemy_Basic prefab.

Fixed the bug where the enemy alerted text was not appearing by modifying the Talk Actions to correct the y position of the text as it was appearing above the sphere collider.