Seventeenth Update
Created a script called P2_EnemyManager to fix the error "Failed to create agent because there is no valid NavMesh" that was occurring in the previous builds of the game.

Recreated the NavMesh surface asset for the Prototype_2_Level scene.

Removed unused variables in the blackboard asset for the P2_EnemyBehaviourGraph behaviour graph.

Updated some packages and removed some unused packages from the package manager.

Duplicated the existing enemy waypoint transforms and added these transforms to the corresponding enemies.

Disabled the Enemy Way Points object and all of its children.

Created a script called P2_FreezeWaypoints to prevent the child waypoint transforms from moving with the enemy parent.

Unpacked all the enemies from the P2_Enemy_Basic prefab and renamed each enemy to "Enemy {integer value}".

Added new child waypoint objects to the Enemy game object.

Updated the OnTriggerEnter function of the P2_Escape_Door script to correctly detect the heist payload game object.

Added a private boolean to the P2_Escape_Door which is updated by the event for when any player character is killed. This changes the requirement in the OnTriggerEnter function to detect the player only.

Removed the reference to the P2_PlayerDetector script in the P2_GameManager script.