Eighteenth Update
Adjusted the waypoints of the four enemies near to the exit door.

Increased the distance threshold in the inspector for the Patrol action node in the P2_EnemyBehaviourGraph.

Updated the collider check for the Heist Payload game object in the OnTriggerEnter function of the P2_Escape_Door script from a tag comparison to a game object name comparison.

Created a new child object of the Heist Payload object called "Heist Object Trigger" and attached a box collider component to it. Checked the trigger option for the box collider.

Changed the damage variable of the P2_GunplayManager script component of the P1 Gun child of Player 1 from 32 to 25.