Nineteenth Update
Fixed a bug where the player gun could hit fire through level geometry and potentially hit enemies.

Added titles to each character name.

Referenced the character name text in the P2_PlayerCharacterBase script and set the text to the character's name in the update function.

Implemented code in the update function of the P2_PlayerManager script so that the canvases of the non-active player characters will look at the active player character.

Fixed a bug where an already hacked door could still be hacked because the trigger collider was still active.

Reduced the range of the P1 gun.

Increase the range of enemies.

Increased the push strength of the Player 1 character to 0.7.

Repositioned each player character's health bar and name text, so that the health bar is below the name text.

Implemented functionality to the P2_PlayerCharacterBase script so that the name text and health of the inactive characters will rotate and look at at the main camera.

Fixed a bug where the enemies would not navigate within range of the detected player character.