Twenty-Third Update
Changed the layer for the MuzzleFlash and SandImpacts prefabs from 'Default' to 'TransparentFX'.

Updated the excluded layers of the trigger box collider of the P2_Hackable_Door prefab to exclude the "Transparent FX" layer.

Disabled the sphere collider of the SandImpacts prefab.

Duplicated the ground level geometry object and renamed it to Ceiling.

Disabled all the components of Ceiling except for the mesh collider.

Added 'Pushable' and 'Environment' to the excluded layers on the mesh collider of the new ceiling game object.

Decreased the wait time for enemies to spawn in on the P2_EnemyManager script component   of the Enemies game object.