Update One
Created a new behaviour graph called P3_BasicEnemyGraph.

Merged the contents of the P3_EnemyHealth script with the P3_EnemyBase script, then deleted the P3_EnemyHealth script.

Fixed the P3_rotatingLights script.

Implemented functionality and events for enemy types and when specific types are killed.

Started working on the P3_LighthouseManager script to implement functionality for repairing and recharging.