Update Five
Updated the BigExplosion prefab to stop it looping after playing once in the scene.

Removed the sphere colliders from the lighthouse game object and its child game object called "Central base."

Changed the Lighthouse Game Object reference in the P3_EnemyWaveManager script component of the EnemyWaveManager from the EnemyDetector game object to the Lighthouse Trigger game object.

Finished implementing and fixing all the functionality for the lighthouse system and corresponding U.I.

Adjusted the movement speed and acceleration values for all the prototype 3 enemies.

Fixed outstanding bugs in the P3_PlayerManager script which were causing issues when switching between the two player characters.

Created two new settings on the setting pages in the pause menu. The first for toggling milliseconds in the survival timer and the second is for toggling bullet trails.