Update Nine
Made further adjustments to the materials for the enemy prefabs.

Fixed a bug where the survival timer would start as soon as the scene loads. This has been fixed so that the timer starts once the player has pressed start game.

Attempted to fix a bug where the repair button would enable when none of the conditions had been met.

Made further adjustments to the lighting.

Reduced the quality of the obstacle avoidance to low in the navmesh agent component of the P3_Enemy_Base prefab.

Installed the adaptive performance package from the project settings.