Update Ten
Made further adjustments to the level lighting, fog and materials for enemy prefabs.

Fixed a bug where pausing the game with the lighthouse player character would not enable the player controls.

Fixed a bug where the audio sliders in the settings page of the pause menu would not move correctly. This was caused by the whole numbers setting being checked for all the corresponding sliders.

Fixed a bug in which the repair button for the lighthouse U.I. would enable.

Fixed the charge limit indicator not showing in the lighthouse U.I. when the lighthouse is damaged.

Fixed a bug where the red type enemies would not move or would not move correctly or could not find the navmesh surface.

Created a new boolean variable for enabling difficulty scaling in the P3_EnemyManager script.

Duplicated the Enemy Wave Manager game object to test enemy spawning.

Disabled the enemy detector game object and the enemy detector light in the scene hierarchy.

Created a new script called P3_NearbyEnemyIndicator. which is for pointing an on screen visual towards the nearest enemy to the player, even if the enemy is off screen.

Created two new child game objects of the Player 1 game object. The first is called Enemy Detector. The second is called Indicator a child of the P1_Camera Position child object.

Created a variant material of the M_Emissive_Red for the indicator game object.

Disabled the Enemy Detector and Indicator child objects of the Player 1 game object as the enemy indicating feature is not yet working.