Update 11
Made further adjustments in the P3_EnemyWaveManager script component of the Enemy Wave Manager game object.

Updated the P3_GameManager script so that all return button functions use the returnSFX audio clip variable.

Created a new game object under canvas called Enemy Information HUD.

Created a new folder under _Source/Art called Prototype 3.

Added enemy types and icon sprites for the Enemy Information HUD to the Prototype 3 art folder.

Set up the Enemy Information HUD with the enemy type and icon sprites.

Added functionality to the P3_GameManager script to toggle the active state of the Enemy Information HUD game object.

Added a new toggle to the settings page of the pause menu to allow the player to toggle the active state of the Enemy Information HUD game object.

Updated the tutorial text.

Updated the text in the P3_LighthouseManager script for number of yellow or blue enemies killed.

Slightly increased the size of the Current Yellow Enemies Killed Text and adjusted its position to accomodate the size increase.

Slightly increased the size of the Current Blue Enemies Killed Text and adjusted its position to accomodate the size increase.

Slightly increased the height of the Recharge Status Text and Repair Status Text objects.

Removed the functionality to toggle the active state of the health bar and health text in the P3_PlayerManager script.

Made adjustments to the speed of the enemy prefabs.

Removed the NavMesh Obstacle component from the Player 1 game object.

Cleared and baked the NavMesh surface.

Updated the quality to good on the Nav Mesh Agent component of all the enemy prefabs.