Various Improvements
Updated the Company Name and Product Name in the Project Settings.

Updated the Scene List in the Windows build profile.

Enabled the main menu buttons in the tutorial and game over screens.

Enabled the main menu button for the tutorial screen in the P3_GameManager script.

Updated the prototype buttons in the main menu.

Added a boolean to the SoundManager script for the main menu, to avoid a null reference error when setting the volume sliders' values in the awake function.

Fixed an issue with the main menu title text not showing.

Updated the prototype 1 main menu button function in the tutorial screen.

Updated the tutorial text for prototype 1.

Scaled up the main menu buttons in the main menu scene.

Changed the background colour of the main menu scene.