Fixes and Minor Features
Updated the Prototype_1_GameManager, P2_GameManager and P3_GameManager scripts to include references to the fog render feature of the URP asset so that the fog is disabled for the first and second prototype scenes.

Added a new text element to the Prototype_1_Level scene called "Enemies Remaining Text".

Added functionality to the Prototype_1_EnemyManager script to update the Enemies Remaining Text.

Fixed the bug where enemy Prototype_1_Projectiles were not hurting the player in the Prototype_1_Level scene.

Fixed an issue where the input action map did not change from the player action map when the player was killed in the Prototype_1_Level scene.

Added outlines to the Health and Ammo bars in the Prototype_1_Level and Prototype_2_Level scenes.