Fixes, Features and Improvements
Added a new bool to the P3_GameManager script called "disableFog" to turn the volumetric fog on and off to improve the shaded lighting mode in scene view.

Added a scroll rect component to the Controls page object in the pause screen of the Prototype_3_Level scene and created a new scrollbar child object called "Scrollbar Vertical.”

Added a scroll rect component to the Controls page object in the pause screen of the Prototype_2_Level scene and created a new scrollbar child object called "Scrollbar Vertical.”

Adjusted the size and position of the controls page contents for both the Prototype_2_Level scene and the Prototype_3_Level scene.

Fixed missing materials for the closed doors of the Prototype_2_Level scene.

Created a new game object under the canvas in the Prototype_2_Level scene called "Player Character Selection Info HUD."

Added functionality to the P2_GameManager script to toggle the active state of the new Player Character Selection Info HUD object.

Added a new toggle to the settings page of the pause menu in the Prototype_2_Level scene to allow the player to toggle the active state of the Player Character Selection Info HUD.

Created a new folder under _Source/Art called Prototype 2.

Added three sprites for each player character in prototype 2.

Updated the P2_PlayerManager script to update the sprites accordingly in the Player Character Selection Info HUD object.

Updated the default text in the P2_Hackable_Door script component of the P2_Hackable_Door prefab.