More Minor Features, Fixes and Improvements
Fixed the enemy indicator in the Prototype_3_Level so it now properly stays on screen and points at enemies properly.

Updated the outlines for all interactive U.I. elements in the settings page of the pause menu for the Prototype_1_Level, Prototype_2_Level and Prototype_3_Level scenes.

Added new settings to the settings page of the pause menu in the Prototype_3_Level scene. The first setting is to toggle the new nearby enemy indicator and the second is a drop down to set the difficulty.

Updated the tutorial text in the Prototype_3_Level scene.

Deleted old text elements from the settings page of the pause screen in the Prototype_1_Level scene.

Fixed a minor bug with the mouse sensitivity settings on the settings page of the pause menu in the Prototype_1_Level scene.

Fixed the OnClick function for the return button of the quit prompt screen in the Prototype_1_Level scene.

Increased the text size of all the pause screen buttons in the Prototype_1_Level scene.