Final Features, Improvements and Bug Fixes
Updated the tutorial text in the Prototype_1_Level scene.

Updated the gun range in the Prototype_1_Level scene.

Updated the attack range for all enemies in the Prototype_1_Level scene.

Updated the level design in the Prototype_2_Level scene to reduce the cube getting stuck in doorways and on corners.

Fixed a bug where a player character would be stuck mid-air if the player switched characters while jumping.

Added sound effects to the P2_HackableDoor script and P2_HackableDoor prefab.

Fixed a bug in the second prototype where switching characters while hacking allowed the player to control two player characters at once.

Fixed a bug in the second prototype where the door was hackable if entering its collider with both the hacking player character and another player character.

Modified the warning screen text in the P2_PlayerDetector script for the Prototype_2_Level scene.

Updated the level design in the Prototype_2_Level scene so to improve pushing and navigating the cube around the level.