Final Fixes
Removed the DataPersistenceManager game object from the MainMenu scene.

Added outlines to the buttons in the of Prototype Selection Page in the MainMenu scene.

Updated the quality levels to high fidelity in the quality settings.

Fixed incorrect links for several sound effects in the credits.

Added an outline to the return button in the credits page of the MainMenu scene.

Updated the button colours for the MainMenu, Prototype_1_Level, Prototype_2_Level and Prototype_3_Level scenes to maintain visual consistency.

Updated the outline colours and sizes for all toggle, slider and button U.I. elements in the MainMenu, Prototype_1_Level, Prototype_2_Level and Prototype_3_Level scenes to maintain visual consistency.

Fixed an issue with enemy sound effects not playing in the Prototype_1_Level scene.

Fixed an issue with removing enemies from the enemies list in the Prototype_1_EnemyManager script component of the enemies game object in the Prototype_1_Level scene.

Fixed an issue with the main menu button in the game over screen in both the Prototype_1_Level and Prototype_2_Level scenes.

Updated the prototype 1 enemy prefabs to make them move slower and to ensure they spawn correctly on the namvesh.

Updated the navmesh surface in the Prototype_1_Level scene.

Updated the level design of the Prototype_1_Level scene to reduce the chances of navmesh agents getting stuck.